Fix lint issues, add desktop-ui react app into eslintrc.js
