+++ .eslintrc.js
@@ -81,6 +81,7 @@ module.exports = {
         'app/**/*.js',
         'shared/**/*.js',
         'ui/**/*.js',
+        'desktop-ui/**/*.js',
         '**/*.test.js',
         'test/lib/**/*.js',
         'test/mocks/**/*.js',
@@ -187,6 +188,7 @@ module.exports = {
         'test/lib/render-helpers.js',
         'test/jest/rendering.js',
         'ui/**/*.js',
+        'desktop-ui/**/*.js',
       ],
       extends: ['plugin:react/recommended', 'plugin:react-hooks/recommended'],
       parserOptions: {
