+++ app/scripts/desktop.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import { render } from 'react-dom';
-import DesktopApp from '../../desktop-ui';
+import launchDesktopUi from '../../desktop-ui';
 
-const rootElement = document.getElementById('root');
+start().catch(console.log);
 
-render(<DesktopApp />, rootElement);
+function start() {
+  launchDesktopUi();
+}
