+++ desktop-ui/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Button from '../ui/components/ui/button';
+
+const App = () => {
+  return (
+    <React.StrictMode>
+      <h1>Desktop app</h1>
+      <Button type="secondary">Button in Desktop</Button>
+    </React.StrictMode>
+  );
+};
+
+export default App;
