+++ package.json
@@ -12,6 +12,7 @@
     "setup:postinstall": "yarn patch-package && yarn allow-scripts",
     "start": "yarn build:dev dev --apply-lavamoat=false",
     "build:desktop": "./development/build-desktop.sh",
+    "build:desktop-ui": "yarn build dist --include-desktop-ui true",
     "build:desktop:mv3": "ENABLE_MV3=true ./development/build-desktop.sh",
     "build:desktop:extension": "DESKTOP=EXTENSION yarn start --build-type flask --include-desktop-ui true",
     "build:desktop:extension:mv3": "DESKTOP=EXTENSION ENABLE_MV3=true yarn start --build-type flask",
@@ -19,7 +20,6 @@
     "start:lavamoat": "yarn build:dev dev --apply-lavamoat=true",
     "start:mv3": "ENABLE_MV3=true yarn build:dev dev --apply-lavamoat=false",
     "dist": "yarn build dist",
-    "dist:desktop-ui": "yarn build dist --include-desktop-ui true",
     "build": "yarn lavamoat:build",
     "build:dev": "node development/build/index.js",
     "start:test": "SEGMENT_HOST='https://api.segment.io' SEGMENT_WRITE_KEY='FAKE' SENTRY_DSN_DEV=https://fake@sentry.io/0000000 PORTFOLIO_URL=http://127.0.0.1:8080 yarn build testDev",
